# PERF: array_equivalent_object