```
from asv_bench.benchmarks.frame_methods import *
self = Equals()
self.setup()

df = self.nonunique_cols
df2 = df.T


%timeit df.equals(df)
26.4 ms ± 667 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
25.9 ms ± 1.36 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR

%timeit df2.equals(df2)
47.3 ms ± 704 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
27.2 ms ± 1.16 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```